test: enable EditorOptionHandlers-integ-test
